Backend/requirements.txt
flask
flask-cors
flask-socketio>=5.3.0
eventlet>=0.33.0
python-socketio>=5.0.0
mysql-connector-python==9.1.0